# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ADMS HR API",
        "description": "API untuk integrasi data absensi antara ADMS dan aplikasi HR eksternal",
        "contact": {
            "email": "admin@adms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
        "/admin/tokens": {
            "get": {
                "tags": [
                    "API Token Management"
                ],
                "summary": "List all API tokens",
                "description": "Retrieve a list of all API tokens (admin only)",
                "operationId": "listApiTokens",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ApiToken"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "API Token Management"
                ],
                "summary": "Generate a new API token",
                "description": "Create a new API token for authentication. The token is only visible in this response.",
                "operationId": "createApiToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Token name/description",
                                        "type": "string",
                                        "example": "HR Application Token"
                                    },
                                    "expires_at": {
                                        "description": "Optional expiration date",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-11-15T00:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Token created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ApiTokenWithToken"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/tokens/{id}/revoke": {
            "put": {
                "tags": [
                    "API Token Management"
                ],
                "summary": "Revoke an API token",
                "description": "Deactivate an API token to prevent further use",
                "operationId": "revokeApiToken",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "API Token ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token revoked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ApiToken"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiToken": {
                "title": "API Token",
                "description": "API Token model (token field is hidden)",
                "properties": {
                    "id": {
                        "description": "Token ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Token name/description",
                        "type": "string",
                        "example": "HR Application Token"
                    },
                    "last_used_at": {
                        "description": "Last time the token was used",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-15T08:30:00Z",
                        "nullable": true
                    },
                    "expires_at": {
                        "description": "Token expiration date",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-11-15T00:00:00Z",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Whether the token is active",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Token creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-15T08:00:00Z"
                    },
                    "updated_at": {
                        "description": "Token last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-15T08:00:00Z"
                    }
                },
                "type": "object"
            },
            "ApiTokenWithToken": {
                "title": "API Token with Token Value",
                "description": "API Token model with token value (only visible on creation)",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiToken"
                    },
                    {
                        "properties": {
                            "token": {
                                "description": "The actual token value (only visible on creation)",
                                "type": "string",
                                "example": "abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response format",
                "properties": {
                    "success": {
                        "description": "Always false for error responses",
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "description": "Error details",
                        "properties": {
                            "code": {
                                "description": "Error code",
                                "type": "string",
                                "example": "INVALID_TOKEN"
                            },
                            "message": {
                                "description": "Human-readable error message",
                                "type": "string",
                                "example": "Token tidak valid atau expired"
                            },
                            "details": {
                                "description": "Additional error context",
                                "type": "object",
                                "example": []
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}